requirements.txt
flask
flask_sqlalchemy
pymongo
requests
gunicorn
flask_login
psycopg2
flask_wtf
